Start working on error linking + massive refactoring
