+++ src/main/java/org/quiltmc/loader/impl/plugin/SolverErrorHelper.java
@@ -16,39 +16,15 @@
 
 package org.quiltmc.loader.impl.plugin;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.PrintWriter;
-import java.io.StringWriter;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Iterator;
-import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-import org.quiltmc.loader.api.ModDependency;
-import org.quiltmc.loader.api.ModDependencyIdentifier;
-import org.quiltmc.loader.api.VersionInterval;
-import org.quiltmc.loader.api.VersionRange;
-import org.quiltmc.loader.api.ModDependency.Only;
-import org.quiltmc.loader.api.gui.QuiltDisplayedError;
-import org.quiltmc.loader.api.gui.QuiltLoaderGui;
-import org.quiltmc.loader.api.gui.QuiltLoaderText;
-import org.quiltmc.loader.api.plugin.ModMetadataExt;
 import org.quiltmc.loader.api.plugin.solver.LoadOption;
 import org.quiltmc.loader.api.plugin.solver.ModLoadOption;
 import org.quiltmc.loader.api.plugin.solver.Rule;
@@ -60,10 +36,27 @@ import org.quiltmc.loader.impl.plugin.quilt.QuiltRuleBreakOnly;
 import org.quiltmc.loader.impl.plugin.quilt.QuiltRuleDep;
 import org.quiltmc.loader.impl.plugin.quilt.QuiltRuleDepAny;
 import org.quiltmc.loader.impl.plugin.quilt.QuiltRuleDepOnly;
-import org.quiltmc.loader.impl.util.FileUtil;
+import org.quiltmc.loader.impl.plugin.solvererror.BreaksAllError;
+import org.quiltmc.loader.impl.plugin.solvererror.BreaksError;
+import org.quiltmc.loader.impl.plugin.solvererror.DependsAnyError;
+import org.quiltmc.loader.impl.plugin.solvererror.DependsError;
+import org.quiltmc.loader.impl.plugin.solvererror.DuplicatesError;
+import org.quiltmc.loader.impl.plugin.solvererror.SolverError;
+import org.quiltmc.loader.impl.plugin.solvererror.UnhandledError;
+import org.quiltmc.loader.impl.plugin.solvererror.graph.Breaks;
+import org.quiltmc.loader.impl.plugin.solvererror.graph.BreaksAll;
+import org.quiltmc.loader.impl.plugin.solvererror.graph.DepLink;
+import org.quiltmc.loader.impl.plugin.solvererror.graph.Depends;
+import org.quiltmc.loader.impl.plugin.solvererror.graph.DependsAny;
+import org.quiltmc.loader.impl.plugin.solvererror.graph.Duplicates;
+import org.quiltmc.loader.impl.plugin.solvererror.graph.Graph;
+import org.quiltmc.loader.impl.plugin.solvererror.graph.Link;
+import org.quiltmc.loader.impl.plugin.solvererror.graph.Mandatory;
+import org.quiltmc.loader.impl.plugin.solvererror.graph.Missing;
+import org.quiltmc.loader.impl.plugin.solvererror.graph.MissingAny;
+import org.quiltmc.loader.impl.plugin.solvererror.graph.Provided;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternal;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternalType;
-import org.quiltmc.loader.impl.util.SystemProperties;
 import org.quiltmc.loader.impl.util.log.Log;
 import org.quiltmc.loader.impl.util.log.LogCategory;
 
@@ -81,16 +74,17 @@ class SolverErrorHelper {
 	/**
 	 * Reports all the errors to the plugin manager.
 	 * <p>
-	 *
+	 * <p>
 	 * If the system property {@value org.quiltmc.loader.impl.util.SystemProperties#PRINT_MOD_SOLVING_ERROR_DOT_GRAPH} is true,
 	 * the dot graph for the errors is logged to the info logger.
 	 */
 	void reportErrors() {
-		if (SystemProperties.getBoolean(SystemProperties.PRINT_MOD_SOLVING_ERROR_DOT_GRAPH, false)) {
-			graph.logGraph();
-		}
+//		if (SystemProperties.getBoolean(SystemProperties.PRINT_MOD_SOLVING_ERROR_DOT_GRAPH, false)) {
+		graph.logGraph();
+//		}
+		SolverErrorReportContext context = new SolverErrorReportContext(this.manager, this.errors, this.graph);
 		for (SolverError error : errors) {
-			error.report(manager);
+			error.report(context);
 		}
 	}
 
@@ -119,7 +113,7 @@ class SolverErrorHelper {
 		graph.clean();
 
 		if (!reportGraphErrors()) {
-			addError(new UnhandledError(graph, rules));
+			addError(new UnhandledError(rules));
 		}
 	}
 
@@ -167,8 +161,9 @@ class SolverErrorHelper {
 					}
 				}
 			} else if (rule instanceof QuiltRuleDepOnly) {
-				Optional<LoadOption> depends = rule.getNodesTo()
-						.stream()
+				Optional<LoadOption> depends = Stream.concat(
+								((QuiltRuleDepOnly) rule).getValidOptions().stream(),
+								((QuiltRuleDepOnly) rule).getWrongOptions().stream())
 						.map(LoadOption.class::cast)
 						.findFirst();
 
@@ -222,20 +217,20 @@ class SolverErrorHelper {
 	private boolean reportGraphErrors() {
 		try {
 			boolean added = false;
-			for (LoadOption option : graph.nodes) {
+			for (LoadOption option : graph.nodes()) {
 				for (Link link : graph.edges(option)) {
 					if (link instanceof Breaks) {
 						Breaks breaks = ((Breaks) link);
 						ModLoadOption from = (ModLoadOption) option;
-						this.addError(new BreaksError(graph, from, breaks.getRule()));
+						this.addError(new BreaksError(from, breaks.getRule()));
 						added = true;
 					} else if (link instanceof BreaksAll) {
 						BreaksAll breaksAll = ((BreaksAll) link);
 						ModLoadOption from = (ModLoadOption) option;
-						if (breaksAll.breaks.size() == 1) {
-							this.addError(new BreaksError(graph, from, breaksAll.getRule().options[0]));
+						if (breaksAll.breaks().size() == 1) {
+							this.addError(new BreaksError(from, breaksAll.getRule().options[0]));
 						} else {
-							this.addError(new BreaksAllError(graph, from, breaksAll.getRule()));
+							this.addError(new BreaksAllError(from, breaksAll.getRule()));
 						}
 						added = true;
 					} else if (link instanceof DepLink) {
@@ -244,9 +239,9 @@ class SolverErrorHelper {
 
 						Set<QuiltRuleDepOnly> rules = flattenUnless(dep.getRule());
 						if (rules.size() > 1) {
-							this.addError(new DependsAnyError(graph, from, rules));
+							this.addError(new DependsAnyError(from, rules));
 						} else {
-							this.addError(new DependsError(graph, from, rules.iterator().next()));
+							this.addError(new DependsError(from, rules.iterator().next()));
 						}
 
 						added = true;
@@ -256,7 +251,7 @@ class SolverErrorHelper {
 				for (Link link : graph.edgesTo(option)) {
 					if (link instanceof Duplicates) {
 						Duplicates duplicates = (Duplicates) link;
-						this.addError(new DuplicatesError(graph, duplicates.id, duplicates.options));
+						this.addError(new DuplicatesError(duplicates.id(), duplicates.options()));
 						added = true;
 					}
 				}
@@ -270,6 +265,7 @@ class SolverErrorHelper {
 
 	/**
 	 * Flattens a {@link QuiltRuleDep} into a set of {@link QuiltRuleDepOnly}s, including all the unless clauses.
+	 *
 	 * @param rule the rule to flatten
 	 * @return a set of all the possible {@link QuiltRuleDepOnly}s
 	 */
@@ -288,1759 +284,5 @@ class SolverErrorHelper {
 		}
 
 		return rules;
-	}
-
-	/**
-	 * Represents a graph of all the rules reported in the errors.
-	 */
-	static class Graph {
-		final Set<LoadOption> nodes = new LinkedHashSet<>();
-		final Map<LoadOption, Set<Link>> edges = new LinkedHashMap<>();
-		final Map<LoadOption, Set<Link>> reverseEdges = new LinkedHashMap<>();
-		final Map<Link, LoadOption> parents = new LinkedHashMap<>();
-
-		/**
-		 * @param option the load option
-		 * @return the dot graph node label for the option
-		 */
-		private static String label(LoadOption option) {
-			if (option instanceof ModLoadOption) {
-				return ((ModLoadOption) option).id() + "\\n" + ((ModLoadOption) option).version();
-			}
-
-			return option.toString();
-		}
-
-
-		/**
-		 * Figures out if a specific load option is mandatory.
-		 * This also returns true for mods provided by mandatory mods.
-		 *
-		 * @param option the load option
-		 * @return {@code true} if the option is a mandatory mod, {@code false} otherwise
-		 */
-		boolean isMandatory(LoadOption option) {
-			boolean mandatory = edgesTo(option).stream().anyMatch(Mandatory.class::isInstance);
-
-			mandatory |= edgesTo(option)
-					.stream()
-					.filter(Provided.class::isInstance)
-					.map(parents::get)
-					.filter(Objects::nonNull)
-					.anyMatch(this::isMandatory);
-
-			return mandatory;
-		}
-
-		/**
-		 * @param option the load option
- 		 * @return {@code true} if the load option is provided by another option, {@code false} otherwise
-		 */
-		boolean isProvided(LoadOption option) {
-			return this.edgesTo(option).stream().anyMatch(Provided.class::isInstance);
-		}
-
-		/**
-		 * Figures out if a specific load option is required.
-		 * If the option is provided, it will also check if the providing option is required.
-		 *
-		 * @param option the load option
-		 * @return {@code true} if the load option is depended on by another option, {@code false} otherwise
-		 */
-		boolean isDepended(LoadOption option) {
-			boolean depended = edgesTo(option).stream().anyMatch(Depends.class::isInstance);
-
-			if (this.isProvided(option)) {
-				depended |= this.edgesTo(option).stream()
-						.filter(Provided.class::isInstance)
-						.map(parents::get)
-						.filter(Objects::nonNull)
-						.anyMatch(this::isDepended);
-			}
-
-			return depended;
-		}
-
-		/**
-		 * Logs the graph to the info logger.
-		 */
-		void logGraph() {
-			StringWriter dotGraphString = new StringWriter();
-			PrintWriter dotGraph = new PrintWriter(dotGraphString);
-			dotGraph.println("digraph G {");
-			dotGraph.println("\troot[style=invis];");
-			dotGraph.println("\tsubgraph cluster_root {");
-			nodes
-					.stream()
-					.filter(this::isMandatory)
-					.forEach(option -> dotGraph.printf("\t\t%s[label=\"%s\", shape=\"Mdiamond\"];\n", option.hashCode(), label(option)));
-			dotGraph.print("\t\tstyle=invis;\n");
-			dotGraph.print("\t}\n");
-			nodes
-					.forEach(option -> {
-						if (this.isMandatory(option)) {
-							dotGraph.printf("\troot->%s[style=invis];\n", option.hashCode());
-						} else {
-							dotGraph.printf("\t%s[label=\"%s\", shape=\"rectangle\"];\n", option.hashCode(), label(option));
-						}
-					});
-
-			nodes.forEach(node -> edges(node).forEach(link -> link.dotGraphEdge(node, dotGraph)));
-			nodes.stream()
-					.flatMap(node -> edgesTo(node).stream().filter(RootLink.class::isInstance))
-					.distinct()
-					.forEach(link -> link.dotGraphEdge(null, dotGraph));
-			dotGraph.println("}");
-			dotGraph.println();
-
-			Log.info(LogCategory.SOLVING, dotGraphString.toString());
-		}
-
-		/**
-		 * Adds a link to the graph.
-		 *
-		 * @param from the parent option
-		 * @param link the link
-		 */
-		public void addLink(@Nullable LoadOption from, Link link) {
-			if (from != null) {
-				edges.computeIfAbsent(from, option -> new LinkedHashSet<>()).add(link);
-				nodes.add(from);
-
-				LoadOption oldParent = parents.put(link, from);
-				if (oldParent != null && !Objects.equals(from, oldParent)) {
-					throw new IllegalStateException("this shouldn't happen");
-				}
-			}
-
-			link.children().forEach(to -> {
-				reverseEdges.computeIfAbsent(to, option -> new HashSet<>()).add(link);
-				nodes.add(to);
-			});
-		}
-
-		/**
-		 * @param from the parent option
-		 * @return the set of links from the option
-		 */
-		public Set<Link> edges(LoadOption from) {
-			return edges.getOrDefault(from, new LinkedHashSet<>());
-		}
-
-		/**
-		 * @param to the child option
-		 * @return the set of links to the option
-		 */
-		public Set<Link> edgesTo(LoadOption to) {
-			return reverseEdges.getOrDefault(to, new LinkedHashSet<>());
-		}
-
-		/**
-		 * Prunes rules that provide no additional information to the graph, such as options from technical load options.
-		 */
-		public void clean() {
-			boolean modified = true;
-			while (modified) {
-				modified = false;
-
-				Set<LoadOption> noLinks = this.parents.values()
-						.stream()
-						.filter(option -> this.edgesTo(option).isEmpty())
-						.collect(Collectors.toCollection(HashSet::new));
-
-				for (LoadOption o : noLinks) {
-					nodes.remove(o);
-					Set<Link> removedLinks = this.edges.remove(o);
-					for (Link link : removedLinks) {
-						this.parents.remove(link);
-					}
-					modified = true;
-				}
-			}
-		}
-	}
-
-	/**
-	 * A link between mod options that indicates some form of relationship.
-	 */
-	interface Link {
-		/**
-		 * @return the children for this load option
-		 */
-		Stream<LoadOption> children();
-
-		/**
-		 * Adds the edge(s) for this link to the dot graph.
-		 *
-		 * @param from the parent for this link
-		 * @param dotGraph the graph output
-		 */
-		default void dotGraphEdge(LoadOption from, PrintWriter dotGraph) {
-			this.children().forEach(child -> {
-				dotGraph.printf("\t%s->%s [label=\"%s\"];\n", from.hashCode(), child.hashCode(), this.getClass().getSimpleName());
-			});
-		}
-	}
-
-	/**
-	 * Represents a link that does not represent a link between two options.
-	 */
-	interface RootLink extends Link {
-	}
-
-	/**
-	 * A link that needs rule associated with it.
-	 * @param <RULE> the type of the rule
-	 */
-	interface RuleLink<RULE extends Rule> extends Link {
-		/**
-		 * @return the rule for this link
-		 */
-		@NotNull RULE getRule();
-	}
-
-	/**
-	 * Represents a breaks that has its requirement.
-	 */
-	static class Breaks implements RuleLink<QuiltRuleBreakOnly> {
-		@NotNull
-		final QuiltRuleBreakOnly rule;
-		@NotNull
-		final LoadOption breaks;
-
-		public Breaks(@NotNull LoadOption breaks, @NotNull QuiltRuleBreakOnly rule) {
-			this.breaks = breaks;
-			this.rule = rule;
-		}
-
-		public Stream<LoadOption> children() {
-			return Stream.of(breaks);
-		}
-
-		@Override
-		public void dotGraphEdge(LoadOption from, PrintWriter dotGraph) {
-			dotGraph.printf("\t%s->%s [label=\"Breaks\", dir=both, color=red];\n", from.hashCode(), breaks.hashCode());
-			if (rule.unless != null) {
-				if (rule.unless instanceof QuiltRuleDepOnly) {
-					for (LoadOption option : rule.unless.getNodesTo()) {
-						dotGraph.printf("\t%s->%s [label=\"Unless\", color=blue];\n", from.hashCode(), option.hashCode());
-					}
-				} else if (rule.unless instanceof QuiltRuleDepAny) {
-					for (QuiltRuleDepOnly only : ((QuiltRuleDepAny) rule.unless).options) {
-						for (LoadOption option : only.getNodesTo()) {
-							dotGraph.printf("\t%s->%s [label=\"Unless\", color=blue];\n", from.hashCode(), option.hashCode());
-						}
-					}
-				}
-			}
-		}
-
-		@Override
-		public @NotNull QuiltRuleBreakOnly getRule() {
-			return rule;
-		}
-
-		@Override
-		public String toString() {
-			return "Breaks=" + breaks;
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) return true;
-			if (o == null || getClass() != o.getClass()) return false;
-
-			Breaks breaks1 = (Breaks) o;
-
-			if (!rule.equals(breaks1.rule)) return false;
-			return breaks.equals(breaks1.breaks);
-		}
-
-		@Override
-		public int hashCode() {
-			int result = rule.hashCode();
-			result = 31 * result + breaks.hashCode();
-			return result;
-		}
-	}
-
-	/**
-	 * Represents a breaks all that is has all of its requirements.
-	 */
-	static class BreaksAll implements RuleLink<QuiltRuleBreakAll> {
-		@NotNull
-		final Collection<LoadOption> breaks;
-		@NotNull
-		final QuiltRuleBreakAll rule;
-
-		public BreaksAll(@NotNull Collection<LoadOption> breaks, @NotNull QuiltRuleBreakAll rule) {
-			this.breaks = breaks;
-			this.rule = rule;
-		}
-
-		public Stream<LoadOption> children() {
-			return breaks.stream();
-		}
-
-		@Override
-		public void dotGraphEdge(LoadOption from, PrintWriter dotGraph) {
-			dotGraph.printf("\t%s->%s [label=\"Breaks\", dir=both, color=red];\n", from.hashCode(), breaks.hashCode());
-			dotGraph.printf("\t%s [label=\"All Of\", shape=\"invtriangle\", color=red];\n", breaks.hashCode());
-			for (LoadOption b : breaks) {
-				dotGraph.printf("\t%s->%s [label=\"Breaks\", dir=both, color=red];\n", breaks.hashCode(), b.hashCode());
-			}
-		}
-
-		@Override
-		public @NotNull QuiltRuleBreakAll getRule() {
-			return this.rule;
-		}
-
-		@Override
-		public String toString() {
-			return "BreaksAll=" + breaks;
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) return true;
-			if (o == null || getClass() != o.getClass()) return false;
-
-			BreaksAll breaksAll = (BreaksAll) o;
-
-			if (!breaks.equals(breaksAll.breaks)) return false;
-			return rule.equals(breaksAll.rule);
-		}
-
-		@Override
-		public int hashCode() {
-			int result = breaks.hashCode();
-			result = 31 * result + rule.hashCode();
-			return result;
-		}
-	}
-
-	/**
-	 * A depend or depends any rule wrapper for code quality.
-	 * @param <RULE> the rule type
-	 */
-	interface DepLink<RULE extends QuiltRuleDep> extends RuleLink<RULE> {
-	}
-
-	/**
-	 * Represents a depend that has its requirement.
-	 */
-	static class Depends implements DepLink<QuiltRuleDepOnly> {
-		@NotNull
-		final QuiltRuleDepOnly rule;
-		@NotNull
-		final LoadOption depends;
-
-		public Depends(@NotNull LoadOption depends, @NotNull QuiltRuleDepOnly rule) {
-			this.depends = depends;
-			this.rule = rule;
-		}
-
-		public Stream<LoadOption> children() {
-			return Stream.of(depends);
-		}
-
-		@Override
-		public @NotNull QuiltRuleDepOnly getRule() {
-			return this.rule;
-		}
-
-		@Override
-		public String toString() {
-			return "Depends=" + depends;
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) return true;
-			if (o == null || getClass() != o.getClass()) return false;
-
-			Depends depends1 = (Depends) o;
-
-			if (!rule.equals(depends1.rule)) return false;
-			return depends.equals(depends1.depends);
-		}
-
-		@Override
-		public int hashCode() {
-			int result = rule.hashCode();
-			result = 31 * result + depends.hashCode();
-			return result;
-		}
-	}
-
-	/**
-	 * Represents a depends any that has some or all of its options.
-	 */
-	static class DependsAny implements DepLink<QuiltRuleDepAny> {
-		@NotNull
-		final Collection<LoadOption> depends;
-
-		final QuiltRuleDepAny rule;
-
-		public DependsAny(@NotNull Collection<LoadOption> depends, QuiltRuleDepAny rule) {
-			this.depends = depends;
-			this.rule = rule;
-		}
-
-		public Stream<LoadOption> children() {
-			return depends.stream();
-		}
-
-		@Override
-		public void dotGraphEdge(LoadOption from, PrintWriter dotGraph) {
-			dotGraph.printf("\t%s->%s [label=\"Depends\"];\n", from.hashCode(), depends.hashCode() + 31);
-			dotGraph.printf("\t%s [label=\"Any Of\", shape=\"invtriangle\"];\n", depends.hashCode() + 31);
-			for (LoadOption b : depends) {
-				dotGraph.printf("\t%s->%s [label=\"Depends\"];\n", depends.hashCode() + 31, b.hashCode());
-			}
-		}
-
-		@Override
-		public @NotNull QuiltRuleDepAny getRule() {
-			return this.rule;
-		}
-
-		@Override
-		public String toString() {
-			return "DependsAny=" + depends;
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) return true;
-			if (o == null || getClass() != o.getClass()) return false;
-
-			DependsAny that = (DependsAny) o;
-
-			return depends.equals(that.depends);
-		}
-
-		@Override
-		public int hashCode() {
-			return depends.hashCode();
-		}
-	}
-
-	/**
-	 * Represents a depend that is missing its requirement.
-	 */
-	static class Missing implements DepLink<QuiltRuleDepOnly> {
-		@NotNull
-		final QuiltRuleDepOnly rule;
-		@NotNull
-		final ModDependency.Only missing;
-
-		public Missing(@NotNull ModDependency.Only missing, @NotNull QuiltRuleDepOnly rule) {
-			this.missing = missing;
-			this.rule = rule;
-		}
-
-		public Stream<LoadOption> children() {
-			return Stream.empty();
-		}
-
-		@Override
-		public @NotNull QuiltRuleDepOnly getRule() {
-			return this.rule;
-		}
-
-		@Override
-		public void dotGraphEdge(LoadOption from, PrintWriter dotGraph) {
-			dotGraph.printf("\t%s->%s [label=\"Depends\"];\n", from.hashCode(), missing.hashCode());
-			dotGraph.printf("\t%s [label=\"Missing: %s\", shape=\"ellipse\", color=red];\n", missing.hashCode(), missing.id());
-		}
-
-		@Override
-		public String toString() {
-			return "Missing=" + missing.id();
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) return true;
-			if (o == null || getClass() != o.getClass()) return false;
-
-			Missing missing1 = (Missing) o;
-
-			if (!rule.equals(missing1.rule)) return false;
-			return missing.equals(missing1.missing);
-		}
-
-		@Override
-		public int hashCode() {
-			int result = rule.hashCode();
-			result = 31 * result + missing.hashCode();
-			return result;
-		}
-	}
-
-	/**
-	 * Represents a depends any that is missing all of its options.
-	 */
-	static class MissingAny implements DepLink<QuiltRuleDepAny> {
-		@NotNull
-		final ModDependency.Any missing;
-
-		final QuiltRuleDepAny rule;
-
-		public MissingAny(@NotNull ModDependency.Any missing, QuiltRuleDepAny rule) {
-			this.missing = missing;
-			this.rule = rule;
-		}
-
-		public Stream<LoadOption> children() {
-			return Stream.empty();
-		}
-
-		@Override
-		public void dotGraphEdge(LoadOption from, PrintWriter dotGraph) {
-			dotGraph.printf("\t%s->%s [label=\"Depends\"];\n", from.hashCode(), missing.hashCode());
-			dotGraph.printf("\t%s [label=\"Missing any of: %s\", shape=\"ellipse\", color=red];\n", missing.hashCode(), missing.stream().map(ModDependency.Only::id).map(ModDependencyIdentifier::toString).collect(Collectors.joining(", ")));
-		}
-
-		@Override
-		public @NotNull QuiltRuleDepAny getRule() {
-			return this.rule;
-		}
-
-		@Override
-		public String toString() {
-			return "MissingAny=[" + missing.stream().map(ModDependency.Only::id).map(ModDependencyIdentifier::toString).collect(Collectors.joining(", ")) + "]";
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) return true;
-			if (o == null || getClass() != o.getClass()) return false;
-
-			MissingAny that = (MissingAny) o;
-
-			return missing.equals(that.missing);
-		}
-
-		@Override
-		public int hashCode() {
-			return missing.hashCode();
-		}
-	}
-
-	/**
-	 * Represents a set of mods that all provide the same id.
-	 */
-	static class Duplicates implements RootLink {
-		@NotNull
-		final Collection<LoadOption> options;
-		@NotNull
-		final String id;
-
-		public Duplicates(@NotNull Collection<LoadOption> options, @NotNull String id) {
-			this.options = options;
-			this.id = id;
-		}
-
-		public Stream<LoadOption> children() {
-			return options.stream();
-		}
-
-		@Override
-		public void dotGraphEdge(LoadOption from, PrintWriter dotGraph) {
-			dotGraph.printf("\t%s[label=\"%s\"];\n", id, id);
-			for (LoadOption option : options) {
-				dotGraph.printf("\t%s->%s [label=\"Duplicates\", style=\"dashed\"];\n", option.hashCode(), id);
-			}
-		}
-
-		@Override
-		public String toString() {
-			return "Duplicates=" + options;
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) return true;
-			if (o == null || getClass() != o.getClass()) return false;
-
-			Duplicates that = (Duplicates) o;
-
-			if (!options.equals(that.options)) return false;
-			return id.equals(that.id);
-		}
-
-		@Override
-		public int hashCode() {
-			int result = options.hashCode();
-			result = 31 * result + id.hashCode();
-			return result;
-		}
-	}
-
-	/**
-	 * Represents a link that a mod is mandatory and must be loaded.
-	 */
-	static class Mandatory implements RootLink {
-		@NotNull
-		final LoadOption option;
-
-		public Mandatory(@NotNull LoadOption option) {
-			this.option = option;
-		}
-
-		public Stream<LoadOption> children() {
-			return Stream.of(option);
-		}
-
-		@Override
-		public void dotGraphEdge(LoadOption from, PrintWriter dotGraph) {
-		}
-
-		@Override
-		public String toString() {
-			return "Manditory=" + option;
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) return true;
-			if (o == null || getClass() != o.getClass()) return false;
-
-			Mandatory mandatory = (Mandatory) o;
-
-			return option.equals(mandatory.option);
-		}
-
-		@Override
-		public int hashCode() {
-			return option.hashCode();
-		}
-	}
-
-	/**
-	 * Represents a link between to options where one provides the other.
-	 */
-	static class Provided implements Link {
-		@NotNull
-		final LoadOption provides;
-
-		public Provided(@NotNull LoadOption provides) {
-			this.provides = provides;
-		}
-
-		public Stream<LoadOption> children() {
-			return Stream.of(provides);
-		}
-
-		@Override
-		public void dotGraphEdge(LoadOption from, PrintWriter dotGraph) {
-			dotGraph.printf("\t%s->%s [label=\"Provides\", style=\"dashed\"];\n", from.hashCode(), provides.hashCode());
-		}
-
-		@Override
-		public String toString() {
-			return "Provided=" + provides;
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) return true;
-			if (o == null || getClass() != o.getClass()) return false;
-
-			Provided provided = (Provided) o;
-
-			return provides.equals(provided.provides);
-		}
-
-		@Override
-		public int hashCode() {
-			return provides.hashCode();
-		}
-	}
-
-	/**
-	 * A solver error which might have multiple real errors merged into one for display.
-	 */
-	static abstract class SolverError {
-		final Graph graph;
-
-		public SolverError(Graph graph) {
-			this.graph = graph;
-		}
-
-		/**
-		 * Attempts to merge this error into the given error. This object itself shouldn't be modified.
-		 *
-		 * @return True if the destination object was modified (and this was merged into it), false otherwise.
-		 */
-		abstract boolean mergeInto(SolverError into);
-
-		/**
-		 * Adds the error to the plugin manager.
-		 *
-		 * @param manager the plugin manager
-		 */
-		abstract void report(QuiltPluginManagerImpl manager);
-
-		/**
-		 * Adds the unless clause to the error.
-		 *
-		 * @param error the error
-		 * @param dep the dependency containing the unless
-		 * @param unlessDep the rule for the unless
-		 */
-		protected void addUnless(QuiltDisplayedError error, ModDependency.Only dep, QuiltRuleDep unlessDep) {
-			if (dep.unless() instanceof ModDependency.Only) {
-				ModDependency.Only unless = (ModDependency.Only) dep.unless();
-				assert unless != null;
-				QuiltLoaderText versionText = VersionRangeDescriber.describe(unless.versionRange(), unless.id().id());
-
-				QuiltRuleDepOnly unlessRule = ((QuiltRuleDepOnly) unlessDep);
-				if (unlessRule.getNodesTo().isEmpty()) {
-					error.appendDescription(QuiltLoaderText.translate("error.break.unless.missing",  versionText));
-				} else {
-					error.appendDescription(QuiltLoaderText.translate("error.break.unless.invalid",  versionText));
-				}
-			} else if (dep.unless() instanceof ModDependency.Any) {
-				ModDependency.Any unless = (ModDependency.Any) dep.unless();
-				assert unless != null;
-
-				error.appendDescription(QuiltLoaderText.translate("error.break.unless_any"));
-
-				for (QuiltRuleDepOnly only : ((QuiltRuleDepAny) unlessDep).options) {
-					QuiltLoaderText versionText = VersionRangeDescriber.describe(only.publicDep.versionRange(), only.publicDep.id().id());
-					if (only.getNodesTo().isEmpty()) {
-						error.appendDescription(QuiltLoaderText.translate("error.break.unless_any.missing",  versionText));
-					} else {
-						error.appendDescription(QuiltLoaderText.translate("error.break.unless_any.invalid",  versionText));
-					}
-				}
-			}
-		}
-
-		/**
-		 * Adds the following mod load options as buttons to the error.
-		 *
-		 * @param error the error
-		 * @param manager the plugin manager to get the paths to the mods
-		 * @param optionsReason the key to the reason for these mods to be mentioned
-		 * @param mods a collection of mods to list
-		 */
-		protected final void addFiles(QuiltDisplayedError error, QuiltPluginManagerImpl manager, String optionsReason, Collection<ModLoadOption> mods) {
-			Map<Path, ModLoadOption> realPaths = new LinkedHashMap<>();
-
-			if (!mods.isEmpty()) {
-				error.appendDescription(QuiltLoaderText.translate(optionsReason));
-			}
-			for (ModLoadOption mod : mods) {
-				boolean provided = graph.isProvided(mod);
-				boolean depended = graph.isDepended(mod);
-				boolean mandatory = graph.isMandatory(mod);
-				// TODO: describe the version?
-				Object[] modDescArgs = {mod.id(), manager.describePath(mod.from())};
-				String key = "info.";
-				if (provided) {
-					key += "provided_";
-				}
-				if (depended) {
-					key += "depended_";
-				}
-				if (!mandatory) {
-					key += "optional_";
-				}
-				key += "mod_loaded_from";
-
-				error.appendDescription(QuiltLoaderText.translate(key, modDescArgs));
-				manager.getRealContainingFile(mod.from()).ifPresent(p -> realPaths.putIfAbsent(p, mod));
-			}
-
-			for (Map.Entry<Path, ModLoadOption> entry : realPaths.entrySet()) {
-				error.addFileViewButton(entry.getKey()).icon(entry.getValue().modCompleteIcon());
-			}
-		}
-
-		/**
-		 * Adds the issue link from the mod to the error.
-		 *
-		 * @param error the error
-		 * @param mandatoryMod the mod
-		 */
-		protected void addIssueLink(QuiltDisplayedError error, ModLoadOption mandatoryMod) {
-			String issuesUrl = mandatoryMod.metadata().contactInfo().get("issues");
-			if (issuesUrl != null) {
-				error.addOpenLinkButton(QuiltLoaderText.translate("button.mod_issue_tracker", mandatoryMod.metadata().name()), issuesUrl);
-			}
-		}
-
-		/**
-		 * Adds the mod icon to the error.
-		 *
-		 * @param error the error
-		 * @param mandatoryMod the mod to get the icon from
-		 */
-		protected void setIconFromMod(QuiltDisplayedError error, ModLoadOption mandatoryMod) {
-			// TODO: Only upload a ModLoadOption's icon once!
-			Map<String, byte[]> images = new HashMap<>();
-			for (int size : new int[]{16, 32}) {
-				String iconPath = mandatoryMod.metadata().icon(size);
-				if (iconPath != null && !images.containsKey(iconPath)) {
-					Path path = mandatoryMod.resourceRoot().resolve(iconPath);
-					try (InputStream stream = Files.newInputStream(path)) {
-						images.put(iconPath, FileUtil.readAllBytes(stream));
-					} catch (IOException io) {
-						Log.error(LogCategory.SOLVING, "Error setting GUI icon for mod %s", mandatoryMod.metadata().name(), io);
-					}
-				}
-			}
-
-			if (!images.isEmpty()) {
-				error.setIcon(QuiltLoaderGui.createIcon(images.values().toArray(new byte[0][])));
-			}
-		}
-
-		/**
-		 * Gets a description for the specified load option, including some relations and the path. It follows the format:
-		 * {@code [[Needed] <Mandatory|Optional> '<mod_name>'] [version '<version>'] [Provided by '<providing_mod_name>'] [, which is contained within '<containing_mod_name>']: <path>}.
-		 *
-		 * @param option     the mod load option
-		 * @param manager    the plugin manager
-		 * @param addName    {@code true} if the name should be added to the mod line. Ignored if the mod is provided
-		 * @param addVersion {@code true} if the version string should be added to the mod line
-		 * @param isDependency {@code true} if the mod is already part of a dependency error
-		 * @return the mod load option description
-		 */
-		protected List<String> getModReportLine(ModLoadOption option, QuiltPluginManagerImpl manager, boolean addName, boolean addVersion, boolean isDependency) {
-			boolean provided = graph.isProvided(option);
-			boolean depended = graph.isDepended(option);
-			boolean optional = !graph.isMandatory(option);
-
-			StringBuilder lineBuilder = new StringBuilder("- ");
-			List<String> lines = new ArrayList<>();
-
-			if (provided) {
-				// This cast is fine because only mod load options can be provided
-				ModLoadOption providingMod = ((ModLoadOption) graph.parents.get(graph.edgesTo(option).stream().filter(Provided.class::isInstance).findFirst().get()));
-				lineBuilder.append("Provided by '").append(providingMod.metadata().name()).append("'");
-			} else {
-				if (!isDependency){
-					if (depended) {
-						lineBuilder.append("Needed");
-					} else if (optional) {
-						lineBuilder.append("Optional");
-					}
-				}
-
-				if (addName) {
-					if ((depended || optional) && !isDependency) {
-						lineBuilder.append(" ");
-					}
-					lineBuilder.append("'").append(option.metadata().name()).append("'");
-				}
-
-				if (addVersion) {
-					if ((depended || optional) && !isDependency || addName) {
-						lineBuilder.append(" v");
-					} else {
-						lineBuilder.append("V");
-					}
-					lineBuilder.append("ersion '").append(option.version()).append("'");
-				}
-			}
-
-			if (option.getContainingMod() != null) {
-				ModLoadOption containingMod = option;
-				List<ModLoadOption> containers = new ArrayList<>();
-				while (containingMod.getContainingMod() != null) {
-					containers.add(containingMod);
-					containingMod = containingMod.getContainingMod();
-				}
-				containers.add(containingMod);
-
-				lineBuilder.append(", which is contained within '").append(containingMod.metadata().name()).append("' ")
-					.append("@ ").append(manager.describePath(containingMod.from())).toString();
-				lines.add(lineBuilder.toString());
-				lineBuilder = new StringBuilder();
-
-				if (containers.size() > 2) {
-					lineBuilder.append("  - Note: '").append(option.metadata().name()).append("' is contained within '").append(containingMod.metadata().name()).append("' through the chain ");
-					Collections.reverse(containers);
-					
-					lineBuilder.append(containers.stream().map(ModLoadOption::metadata).map(ModMetadataExt::name)
-						.map(name -> "'" + name + "'")
-						.collect(Collectors.joining(" -> ")));
-					lines.add(lineBuilder.toString());
-				}
-			} else {
-				if (provided || addName || addVersion) {
-					lineBuilder.append(" ");
-				}
-
-				String line = lineBuilder.append("@ ").append(manager.describePath(option.from())).toString();
-				lines.add(line);
-			}
-			return lines;
-		}
-
-		/**
-		 * Adds the version string for the dependency to the string builder, and whitespace is added around the string.
-		 * If {@code start} is true, this is expected to be the start of a new line, and no leading space is added.
-		 *
-		 * @param report the string builder to inline the string into
-		 * @param only the dependency
-		 * @param breaks {@code true} if this is a breaking dependency, {@code false} if it is a required dependency
-		 * @param start {@code true} if this is the start of a line, {@code false} otherwise.
-		 */
-		protected void addVersionString(StringBuilder report, Only only, boolean breaks, boolean start) {
-			StringBuilder version = new StringBuilder();
-			if (breaks) {
-				version.append(" all versions");
-			} else {
-				version.append(" any version");
-			}
-
-			if (only.versionRange().size() == 1) {
-				VersionInterval interval = only.versionRange().getFirst();
-
-				if (interval.getMin() == null) { // neg infinity
-					if (interval.getMax() == null) {
-						// any
-						version.append(" of ");
-					} else {
-						if (interval.isMaxInclusive()) {
-							version.append(" lesser than or equal to ")
-								.append(interval.getMax())
-								.append(" of ");
-						} else {
-							version.append(" lesser than ")
-								.append(interval.getMax())
-								.append(" of ");
-						}
-					}
-				} else {
-					if (interval.getMax() == null) { // pos infinity
-						if (interval.isMinInclusive()) {
-							version.append(" greater than or equal to ")
-								.append(interval.getMin())
-								.append(" of ");
-						} else {
-							version.append(" greater than ")
-								.append(interval.getMin())
-								.append(" of ");
-						}
-					} else {
-						if (interval.getMin().equals(interval.getMax())) {
-							version = new StringBuilder(" exactly version ")
-								.append(interval.getMin())
-								.append(" of ");
-						} else {
-							if (interval.isMinInclusive()) {
-								version.append(" >=");
-							} else {
-								version.append(" >");
-							}
-
-							version.append(interval.getMin()).append(" and ");
-
-							if (interval.isMaxInclusive()) {
-								version.append("<=");
-							} else {
-								version.append("<");
-							}
-
-							version.append(interval.getMax()).append(" of ");
-						}
-					}
-				}
-			} else {
-				version = new StringBuilder(" a version ").append(only.versionRange()).append(" of ");
-			}
-
-			String versionString = version.toString();
-			if (start) {
-				char second = versionString.charAt(1);
-				versionString = Character.toUpperCase(second) + versionString.substring(2);
-			}
-
-			report.append(versionString);
-		}
-
-		/**
-		 * Describes the unless clause for a breaking dependency. We do not handle a requiring dependency, since those algebraically resolve to a DepAny.
-		 *
-		 * @param dep the dependency containing the unless
-		 * @param dependentName the name of the dependent mod(s)
-		 * @param unlessDepRule the rule specifying the unless
-		 * @param singleDependent {@code true} if there is only one dependent mod, {@code false} if there are multiple
-		 *
-		 * @return the lines describing the unless clause
-		 */
-		protected List<String> addUnlessClause(Only dep, String dependentName, QuiltRuleDep unlessDepRule, boolean singleDependent) {
-			StringBuilder report = new StringBuilder();
-			if (dep.unless() instanceof ModDependency.Only) {
-				ModDependency.Only unless = (ModDependency.Only) dep.unless();
-				assert unless != null;
-				QuiltRuleDepOnly unlessRule = ((QuiltRuleDepOnly) unlessDepRule);
-				if (unlessRule.getNodesTo().isEmpty()) {
-					report.append("However, if");
-
-					addVersionString(report, unless, false, false);
-
-					report.append(unless.id()).append(" is present, ").append(dependentName).append(" do");
-					if (singleDependent) {
-						report.append("es");
-					}
-					report.append(" not break ");
-
-					report.append(dep.id()).append(".");
-				} else {
-					report.append("Normally,");
-
-					addVersionString(report, unless, false, false);
-
-					report.append("mod ").append(unless.id()).append(" overrides this break, but it is unable to load due to another error.");
-				}
-
-				return Collections.singletonList(report.toString());
-			} else if (dep.unless() instanceof ModDependency.Any) {
-				ModDependency.Any unless = (ModDependency.Any) dep.unless();
-				assert unless != null;
-				QuiltRuleDepAny unlessRule = ((QuiltRuleDepAny) unlessDepRule);
-				List<String> lines = new ArrayList<>();
-				report.append("However, if any of the following are present, ");
-				report.append(dependentName).append(" do");
-				if (singleDependent) {
-					report.append("es");
-				}
-				report.append(" not break.");
-				lines.add(report.toString());
-
-				for (QuiltRuleDepOnly only : unlessRule.options) {
-					report = new StringBuilder("- ");
-					addVersionString(report, only.publicDep, false, true);
-					report.append(only.publicDep.id());
-
-					if (only.getAllOptions().isEmpty()) {
-						report.append(" which is missing.");
-					} else {
-						report.append(" which is unable to load due to another error.");
-					}
-					lines.add(report.toString());
-				}
-				return lines;
-			}
-			return Collections.emptyList();
-		}
-	}
-
-	/**
-	 * Reports a set of rules that cause an unknown error.
-	 */
-	static class UnhandledError extends SolverError {
-
-		final Collection<Rule> rules;
-
-		public UnhandledError(Graph graph, Collection<Rule> rules) {
-			super(graph);
-			this.rules = new LinkedHashSet<>(rules);
-		}
-
-		@Override
-		boolean mergeInto(SolverError into) {
-			if (into instanceof UnhandledError) {
-				UnhandledError depDst = (UnhandledError) into;
-				if (depDst.rules.containsAll(rules)) { // If the other error has all the same rules as us, we are just a subset
-					return true;
-				} else if (rules.containsAll(depDst.rules)) { // We have the same but more rules than the current error, add ours to it
-					depDst.rules.addAll(rules);
-					return true;
-				}
-			}
-			return false;
-		}
-
-		@Override
-		void report(QuiltPluginManagerImpl manager) {
-			QuiltDisplayedError error = manager.theQuiltPluginContext.reportError(
-					QuiltLoaderText.translate("error.unhandled_solver")
-			);
-			error.appendDescription(QuiltLoaderText.translate("error.unhandled_solver.desc"));
-			error.addOpenQuiltSupportButton();
-			error.appendReportText("Unhandled solver error involving the following rules:");
-
-			StringBuilder sb = new StringBuilder();
-			int number = 1;
-			for (Rule rule : rules) {
-				error.appendDescription(QuiltLoaderText.translate("error.unhandled_solver.desc.rule_n", number, rule.getClass()));
-				rule.appendRuleDescription(error::appendDescription);
-				error.appendReportText("Rule " + number++ + ":");
-				sb.setLength(0);
-				// TODO: Rename 'fallbackErrorDescription'
-				// to something like 'fallbackReportDescription'
-				// and then clean up all of the implementations to be more readable.
-				rule.fallbackErrorDescription(sb);
-				error.appendReportText(sb.toString());
-			}
-			error.appendReportText("");
-		}
-	}
-
-	/**
-	 * Reports on a set of rules that indicate an issue with a required dependency.
-	 * This can either be that it is missing, or that it is unable to load do to another error.
-	 */
-	static class DependsError extends SolverError {
-		final Set<ModLoadOption> from = new LinkedHashSet<>();
-		final QuiltRuleDepOnly depends;
-
-		DependsError(Graph graph, ModLoadOption from, QuiltRuleDepOnly depends) {
-			super(graph);
-			this.from.add(from);
-			this.depends = depends;
-		}
-
-		@Override
-		boolean mergeInto(SolverError into) {
-			if (into instanceof DependsError) {
-				DependsError depDst = (DependsError) into;
-				if (!depends.publicDep.id().equals(depDst.depends.publicDep.id()) || !depends.publicDep.versionRange().equals(depDst.depends.publicDep.versionRange())) {
-					return false;
-				}
-				depDst.from.addAll(from);
-				return true;
-			}
-			return false;
-		}
-
-		@Override
-		void report(QuiltPluginManagerImpl manager) {
-
-			// Title:
-			// "BuildCraft" requires [version 1.5.1] of "Quilt Standard Libraries", which is
-			// missing!
-
-			// Description:
-			// BuildCraft is loaded from '<mods>/buildcraft-9.0.0.jar'
-			ModLoadOption mandatoryMod = from.iterator().next();
-			String rootModName = from.size() > 1 ? from.size() + " mods [" + from.stream().map(ModLoadOption::metadata).map(ModMetadataExt::name).map(name -> "'" + name + "'").collect(Collectors.joining(", ")) + "]" : ("'" + mandatoryMod.metadata().name() + "'");
-
-			VersionRange range = depends.publicDep.versionRange();
-			String depName = depends.publicDep.id().id();
-			QuiltLoaderText first = VersionRangeDescriber.describe(rootModName, range, depName, true, false);
-
-			Object[] secondData = new Object[(depends.getWrongOptions().size() == 1 || depends.getValidOptions().size() == 1) ? 1 : 0];
-			String secondKey = "error.dep.";
-			if (depends.getAllOptions().isEmpty()) {
-				secondKey += "missing";
-			} else if (depends.getWrongOptions().size() > 1) {
-				secondKey += "multi_mismatch";
-			} else if (depends.getWrongOptions().size() == 1) {
-				secondKey += "single_mismatch";
-				secondData[0] = depends.getWrongOptions().iterator().next().version().toString();
-			} else if (depends.getValidOptions().size() > 1) {
-				secondKey += "multi_valid";
-			} else {
-				secondKey += "single_valid";
-				secondData[0] = depends.getValidOptions().iterator().next().version().toString();
-			}
-			QuiltLoaderText second = QuiltLoaderText.translate(secondKey + ".title", secondData);
-			QuiltLoaderText title = QuiltLoaderText.translate("error.dep.join.title", first, second);
-			QuiltDisplayedError error = manager.theQuiltPluginContext.reportError(title);
-
-			setIconFromMod(error, mandatoryMod);
-			addFiles(error, manager, "error.dep.requiring", from);
-			addFiles(error, manager, "error.dep.valid", depends.getValidOptions());
-			addFiles(error, manager, "error.dep.invalid", depends.getWrongOptions());
-
-			addIssueLink(error, mandatoryMod);
-
-			StringBuilder report = new StringBuilder("Failed to load ")
-				.append(rootModName)
-				.append(" because ")
-				.append(from.size() > 1 ? "they" : "it")
-				.append(" needs");
-			if (!depends.publicDep.versionRange().equals(VersionRange.ANY)) {
-				addVersionString(report, depends.publicDep, false, false);
-			} else {
-				report.append(" ");
-			}
-			report.append("'");
-			String depModName = depends.getAllOptions()
-				.stream()
-				.map(ModLoadOption::metadata)
-				.map(ModMetadataExt::name)
-				.findAny()
-				.orElseGet(depends.publicDep.id()::toString);
-			report.append(
-				depModName
-			).append("'");
-			if (!depends.getValidOptions().isEmpty()) {
-				report.append(", which is unable to load due to another error!");
-			} else if (depends.getWrongOptions().isEmpty()) {
-				report.append(", which is missing!");
-			} else {
-				// Log an error here?
-				report.append(".");
-			}
-
-			error.appendReportText(report.toString(), "");
-
-			if (!depends.publicDep.reason().isEmpty()) {
-				error.appendReportText("The reason " + rootModName + " needs '" + depModName + "': " + depends.publicDep.reason());
-			}
-
-			error.appendReportText("Mod files that need '" + depModName + "': ");
-			for (ModLoadOption mod : from) {
-				getModReportLine(mod, manager, true, false, false).forEach(error::appendReportText);
-			}
-
-			if (!depends.getValidOptions().isEmpty()) {
-				error.appendReportText("");
-				error.appendReportText("Valid mods: ");
-				for (ModLoadOption mod : depends.getValidOptions()) {
-					getModReportLine(mod, manager, true, !depends.publicDep.versionRange().equals(VersionRange.ANY), true).forEach(error::appendReportText);
-					// TODO: link other errors here.
-				}
-			}
-
-			if (!depends.getWrongOptions().isEmpty()) {
-				error.appendReportText("");
-				error.appendReportText("Mods with incorrect versions: ");
-				for (ModLoadOption mod : depends.getWrongOptions()) {
-					getModReportLine(mod, manager, true, !depends.publicDep.versionRange().equals(VersionRange.ANY), false).forEach(error::appendReportText);
-				}
-			}
-		}
-	}
-
-	/**
-	 * Reports on a set of rules that indicate an issue with a required any dependency.
-	 * It handles each only dep as either missing or unable to load.
-	 */
-	static class DependsAnyError extends SolverError {
-		final Set<ModLoadOption> from = new LinkedHashSet<>();
-		final Set<QuiltRuleDepOnly> depends;
-
-		DependsAnyError(Graph graph, ModLoadOption from, Set<QuiltRuleDepOnly> dependsAll) {
-			super(graph);
-			this.from.add(from);
-			this.depends = dependsAll;
-		}
-
-		@Override
-		boolean mergeInto(SolverError into) {
-			if (into instanceof DependsAnyError) {
-				DependsAnyError depDst = (DependsAnyError) into;
-				Set<ModDependency.Only> ours = this.depends.stream().map(rule -> rule.publicDep).collect(Collectors.toSet());
-				Set<ModDependency.Only> theirs = depDst.depends.stream().map(rule -> rule.publicDep).collect(Collectors.toSet());
-
-				if (ours.equals(theirs)) {
-					depDst.from.addAll(from);
-					return true;
-				}
-			}
-			return false;
-		}
-
-		@Override
-		void report(QuiltPluginManagerImpl manager) {
-
-			// Title:
-			// "BuildCraft" depends on any of Quilt Standard Libraries and Minecraft!
-			// "BuildCraft" depends on any of Quilt Standard Libraries, Minecraft, and Third Mod!
-
-			// Description:
-			// BuildCraft is loaded from '<mods>/buildcraft-9.0.0.jar'
-			ModLoadOption mandatoryMod = from.iterator().next();
-			String rootModName = from.size() > 1 ? from.size() + " mods [" + from.stream().map(ModLoadOption::metadata).map(ModMetadataExt::name).collect(Collectors.joining(", ")) + "]" : mandatoryMod.metadata().name();
-
-			Iterator<Only> depends = this.depends.stream().map(rule -> rule.publicDep).iterator();
-			QuiltLoaderText allMods = QuiltLoaderText.of(depends.next().id().id());
-			while (depends.hasNext()) {
-				Only next = depends.next();
-
-				if (depends.hasNext()) {
-					allMods = QuiltLoaderText.translate("error.dep.join.title", allMods, next.id().id());
-				} else {
-					allMods = QuiltLoaderText.translate("error.dep.join.last.title", allMods, next.id().id());
-				}
-			}
-
-			QuiltLoaderText title = QuiltLoaderText.translate("error.dep_any.title", rootModName, allMods);
-			QuiltDisplayedError error = manager.theQuiltPluginContext.reportError(title);
-
-			setIconFromMod(error, mandatoryMod);
-
-			for (QuiltRuleDepOnly depOnly: this.depends) {
-				error.appendDescription(VersionRangeDescriber.describe(depOnly.publicDep.versionRange(), depOnly.publicDep.id().id()));
-				if (!depOnly.publicDep.reason().isEmpty()) {
-					error.appendDescription(QuiltLoaderText.translate("error.reason.specific", depOnly.publicDep.id().id(), depOnly.publicDep.reason()));
-				}
-
-				addFiles(error, manager, "error.dep.valid", depOnly.getValidOptions());
-				addFiles(error, manager, "error.dep.invalid", depOnly.getWrongOptions());
-				error.appendDescription(QuiltLoaderText.of(""));
-			}
-
-			addFiles(error, manager, "error.dep.requiring", from);
-
-			addIssueLink(error, mandatoryMod);
-
-			StringBuilder report = new StringBuilder(rootModName);
-			report.append(" depend");
-			if (from.size() == 1) {
-				report.append("s");
-			}
-			report.append(" on any of the following mods:");
-
-			boolean skippedBreak = false;
-
-			error.appendReportText(report.toString());
-			for (QuiltRuleDepOnly depOnly: this.depends) {
-				skippedBreak = false;
-
-				report = new StringBuilder("- ");
-				addVersionString(report, depOnly.publicDep, false, true);
-				report.append(depOnly.publicDep.id()); // TODO
-
-				if (depOnly.getAllOptions().isEmpty()) {
-					report.append(", which is missing!");
-					if (depOnly.publicDep.reason().isEmpty()) {
-						error.appendReportText(report.toString());
-						skippedBreak = true;
-						continue;
-					}
-				} else {
-					report.append(":");
-				}
-
-				error.appendReportText(report.toString());
-				if (!depOnly.publicDep.reason().isEmpty()) {
-					error.appendReportText("  Depend reason: " + depOnly.publicDep.reason());
-				}
-
-				if (!depOnly.getValidOptions().isEmpty()) {
-					error.appendReportText("  Satisfying mods which cannot load: ");
-					for (ModLoadOption mod : depOnly.getValidOptions()) {
-						error.appendReportText("  " + getModReportLine(mod, manager, true, !depOnly.publicDep.versionRange().equals(VersionRange.ANY), false));
-					}
-				}
-
-				if (!depOnly.getWrongOptions().isEmpty()) {
-					error.appendReportText("  Invalid mods: ");
-					for (ModLoadOption mod : depOnly.getWrongOptions()) {
-						error.appendReportText("  " + getModReportLine(mod, manager, true, !depOnly.publicDep.versionRange().equals(VersionRange.ANY), false));
-					}
-				}
-				error.appendReportText("");
-			}
-
-			if (skippedBreak) {
-				error.appendReportText("");
-			}
-
-			error.appendReportText("Requiring mods: ");
-			for (ModLoadOption mod : from) {
-				getModReportLine(mod, manager, true, false, false).forEach(error::appendReportText);
-			}
-		}
-	}
-
-	/**
-	 * Reports on a set of rules that indicate an issue with a breaking dependency.
-	 */
-	static class BreaksError extends SolverError {
-		final Set<ModLoadOption> from = new LinkedHashSet<>();
-		final QuiltRuleBreakOnly breakage;
-
-		BreaksError(Graph graph, ModLoadOption from, QuiltRuleBreakOnly breakage) {
-			super(graph);
-			this.from.add(from);
-			this.breakage = breakage;
-		}
-
-		@Override
-		boolean mergeInto(SolverError into) {
-			if (into instanceof BreaksError) {
-				BreaksError depDst = (BreaksError) into;
-				if (!breakage.publicDep.id().equals(depDst.breakage.publicDep.id()) || !breakage.publicDep.versionRange().equals(depDst.breakage.publicDep.versionRange())) {
-					return false;
-				}
-				depDst.from.addAll(from);
-				return true;
-			}
-			return false;
-		}
-
-		@Override
-		void report(QuiltPluginManagerImpl manager) {
-			// Title:
-			// "BuildCraft" breaks with [version 1.5.1] of "Quilt Standard Libraries", but it's present!
-
-			// Description:
-			// BuildCraft is loaded from '<mods>/buildcraft-9.0.0.jar'
-			ModLoadOption mandatoryMod = from.iterator().next();
-			String rootModName = from.size() > 1 ? from.size() + " mods [" + from.stream().map(ModLoadOption::metadata).map(ModMetadataExt::name).collect(Collectors.joining(", ")) + "]" : mandatoryMod.metadata().name();
-
-			QuiltLoaderText first = VersionRangeDescriber.describe(rootModName, breakage.publicDep.versionRange(), breakage.publicDep.id().id(), false, false);
-
-			Object[] secondData = new Object[breakage.getConflictingOptions().size() == 1 ? 1 : 0];
-			String secondKey = "error.break.";
-			if (breakage.getConflictingOptions().size() > 1) {
-				secondKey += "multi_conflict";
-			} else {
-				secondKey += "single_conflict";
-				secondData[0] = breakage.getConflictingOptions().iterator().next().version().toString();
-			}
-			QuiltLoaderText second = QuiltLoaderText.translate(secondKey + ".title", secondData);
-			QuiltLoaderText title = QuiltLoaderText.translate("error.break.join.title", first, second);
-
-			if (breakage.publicDep.versionRange().equals(VersionRange.ANY)) {
-				title = QuiltLoaderText.translate(secondKey + ".all.title", rootModName, breakage.publicDep.id().id());
-			}
-
-			QuiltDisplayedError error = manager.theQuiltPluginContext.reportError(title);
-
-			setIconFromMod(error, mandatoryMod);
-
-			if (!breakage.publicDep.reason().isEmpty()) {
-				error.appendDescription(QuiltLoaderText.translate("error.reason", breakage.publicDep.reason()));
-			}
-
-			if (breakage.publicDep.unless() != null) {
-				addUnless(error, breakage.publicDep, breakage.unless);
-				// A newline after the reason was desired here, but do you think Swing loves nice things?
-			}
-
-			addFiles(error, manager, "error.break.breaking", from);
-			addFiles(error, manager, "error.break.broken", breakage.getConflictingOptions());
-			addIssueLink(error, mandatoryMod);
-
-			StringBuilder report = new StringBuilder(rootModName);
-			report.append(" break");
-			if (from.size() == 1) {
-				report.append("s");
-			}
-			addVersionString(report, breakage.publicDep, true, false);
-			report.append(breakage.publicDep.id());// TODO
-			report.append(", which is present!");
-			error.appendReportText(report.toString());
-			if (breakage.publicDep.unless() != null) {
-				addUnlessClause(breakage.publicDep, rootModName, breakage.unless, from.size() == 1).forEach(error::appendReportText);
-			}
-			if (!breakage.publicDep.reason().isEmpty()) {
-				error.appendReportText("Breaking mod's reason: " + breakage.publicDep.reason(), "");
-			}
-			error.appendReportText("");
-
-			error.appendReportText("Breaking mods: ");
-			for (ModLoadOption mod : from) {
-				getModReportLine(mod, manager, true, false, false).forEach(error::appendReportText);
-			}
-
-			error.appendReportText("", "Broken mods: ");
-			for (ModLoadOption mod : breakage.getConflictingOptions()) {
-				getModReportLine(mod, manager, true, !breakage.publicDep.versionRange().equals(VersionRange.ANY), false).forEach(error::appendReportText);
-			}
-
-			if (breakage.unless != null) {
-				if (breakage.unless instanceof QuiltRuleDepOnly && !breakage.unless.getNodesTo().isEmpty()) {
-					error.appendReportText("", "Overriding mods: ");
-					for (LoadOption option : breakage.unless.getNodesTo()) {
-						if (option instanceof ModLoadOption) {
-							getModReportLine(((ModLoadOption) option), manager, true, true, false).forEach(error::appendReportText);
-						} else {
-							error.appendReportText("- " + option.describe().toString());
-						}
-					}
-				} else if (breakage.unless instanceof QuiltRuleDepAny) {
-					boolean added = false;
-					for (QuiltRuleDepOnly only: ((QuiltRuleDepAny) breakage.unless).options) {
-						for (LoadOption option : only.getNodesTo()) {
-							if (!added) {
-								error.appendReportText("", "Overriding mods: ");
-								added = true;
-							}
-							if (option instanceof ModLoadOption) {
-								getModReportLine(((ModLoadOption) option), manager, true, true, false).forEach(error::appendReportText);
-							} else {
-								error.appendReportText("- " + option.describe().toString());
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	/**
-	 * Reports on a set of rules that indicate an issue with a breaks all dependency.
-	 */
-	static class BreaksAllError extends SolverError {
-		final Set<ModLoadOption> from = new LinkedHashSet<>();
-		final QuiltRuleBreakAll breakage;
-
-		BreaksAllError(Graph graph, ModLoadOption from, QuiltRuleBreakAll breakage) {
-			super(graph);
-			this.from.add(from);
-			this.breakage = breakage;
-		}
-
-		@Override
-		boolean mergeInto(SolverError into) {
-			// BreaksAll are rare and there probably wont be duplicates.
-			// Plus collection equals are difficult
-			// We will just make sure we dont have multiple errors for the same rule
-
-			if (into instanceof BreaksAllError) {
-				BreaksAllError depDst = (BreaksAllError) into;
-				return breakage == depDst.breakage;
-			}
-			return false;
-		}
-
-		@Override
-		void report(QuiltPluginManagerImpl manager) {
-
-			// Title:
-			// "BuildCraft" breaks because all of Quilt Standard Libraries and Minecraft are present!
-			// "BuildCraft" breaks because all of Quilt Standard Libraries, Minecraft, and Third Mod are present!
-
-			// Description:
-			// BuildCraft is loaded from '<mods>/buildcraft-9.0.0.jar'
-			ModLoadOption mandatoryMod = from.iterator().next();
-			String rootModName = from.size() > 1 ? from.size() + " mods [" + from.stream().map(ModLoadOption::metadata).map(ModMetadataExt::name).collect(Collectors.joining(", ")) + "]" : mandatoryMod.metadata().name();
-
-			Iterator<Only> breaks = breakage.publicDep.iterator();
-			QuiltLoaderText allMods = QuiltLoaderText.of(breaks.next().id().id());
-			while (breaks.hasNext()) {
-				Only next = breaks.next();
-
-				if (breaks.hasNext()) {
-					allMods = QuiltLoaderText.translate("error.break.join.title", allMods, next.id().id());
-				} else {
-					allMods = QuiltLoaderText.translate("error.break.join.last.title", allMods, next.id().id());
-				}
-			}
-
-			QuiltLoaderText title = QuiltLoaderText.translate("error.breaks_all.title", rootModName, allMods);
-			QuiltDisplayedError error = manager.theQuiltPluginContext.reportError(title);
-
-			setIconFromMod(error, mandatoryMod);
-
-			for (QuiltRuleBreakOnly breakOnly: breakage.options) {
-				error.appendDescription(VersionRangeDescriber.describe(breakOnly.publicDep.versionRange(), breakOnly.publicDep.id().id()));
-				if (!breakOnly.publicDep.reason().isEmpty()) {
-					error.appendDescription(QuiltLoaderText.translate("error.reason.specific", breakOnly.publicDep.id().id(), breakOnly.publicDep.reason()));
-				}
-
-				if (breakOnly.publicDep.unless() != null) {
-					// [VERSION of MOD] can override this break if present.
-					// [VERSION of MOD] overrides this break, but is unable to load due to another error.
-					addUnless(error, breakOnly.publicDep, breakOnly.unless);
-				}
-
-				addFiles(error, manager, "error.break.broken", breakOnly.getConflictingOptions());
-				error.appendDescription(QuiltLoaderText.of(""));
-			}
-
-			
-			addFiles(error, manager, "error.break.breaking", from);
-			
-			addIssueLink(error, mandatoryMod);
-
-			StringBuilder report = new StringBuilder(rootModName);
-			report.append(" break");
-			if (from.size() == 1) {
-				report.append("s");
-			}
-			report.append(" because all of the following are present:");
-			error.appendReportText(report.toString());
-			for (QuiltRuleBreakOnly breakOnly: breakage.options) {
-				report = new StringBuilder("- ");
-				addVersionString(report, breakOnly.publicDep, false, true);
-				report.append(breakOnly.publicDep.id()); // TODO
-				report.append(":");
-				error.appendReportText(report.toString());
-				if (breakOnly.publicDep.unless() != null) {
-					addUnlessClause(breakOnly.publicDep, rootModName, breakOnly.unless, from.size() == 1).forEach(line -> error.appendReportText("  " + line));
-				}
-				if (!breakOnly.publicDep.reason().isEmpty()) {
-					error.appendReportText("  Breaking reason: " + breakOnly.publicDep.reason());
-				}
-	
-				error.appendReportText("  Matching mods: ");
-				for (ModLoadOption mod : breakOnly.getConflictingOptions()) {
-					error.appendReportText("  " + getModReportLine(mod, manager, true, !breakOnly.publicDep.versionRange().equals(VersionRange.ANY), false));
-				}
-	
-				if (breakOnly.unless != null) {
-					if (breakOnly.unless instanceof QuiltRuleDepOnly && !breakOnly.unless.getNodesTo().isEmpty()) {
-						error.appendReportText("  Overriding mods: ");
-						for (LoadOption option : breakOnly.unless.getNodesTo()) {
-							if (option instanceof ModLoadOption) {
-								error.appendReportText("  " + getModReportLine(((ModLoadOption) option), manager, true, !((QuiltRuleDepOnly) breakOnly.unless).publicDep.versionRange().equals(VersionRange.ANY), false));
-							} else {
-								error.appendReportText("  - " + option.describe().toString());
-							}
-						}
-					} else if (breakOnly.unless instanceof QuiltRuleDepAny) {
-						boolean added = false;
-						for (QuiltRuleDepOnly only: ((QuiltRuleDepAny) breakOnly.unless).options) {
-							for (LoadOption option : only.getNodesTo()) {
-								if (!added) {
-									error.appendReportText("  Overriding mods: ");
-									added = true;
-								}
-								if (option instanceof ModLoadOption) {
-									error.appendReportText("  " + getModReportLine(((ModLoadOption) option), manager, true, !only.publicDep.versionRange().equals(VersionRange.ANY), false));
-								} else {
-									error.appendReportText("  - " + option.describe().toString());
-								}
-							}
-						}
-					}
-				}
-				error.appendReportText("");
-			}
-
-			error.appendReportText("Breaking mods: ");
-			for (ModLoadOption mod : from) {
-				getModReportLine(mod, manager, true, false, false).forEach(error::appendReportText);
-			}
-		}
-	}
-
-	/**
-	 * Reports on a set of rules that indicate multiple mods provide the same id.
-	 */
-	static class DuplicatesError extends SolverError {
-		final String id;
-		final Set<LoadOption> duplicates = new LinkedHashSet<>();
-
-		DuplicatesError(Graph graph, String id, Collection<LoadOption> duplicates) {
-			super(graph);
-			this.id = id;
-			this.duplicates.addAll(duplicates);
-		}
-
-		@Override
-		boolean mergeInto(SolverError into) {
-			if (into instanceof DuplicatesError) {
-				DuplicatesError depDst = (DuplicatesError) into;
-				if (!this.id.equals(depDst.id)) {
-					return false;
-				}
-				depDst.duplicates.addAll(duplicates);
-				return true;
-			}
-			return false;
-		}
-
-		@Override
-		void report(QuiltPluginManagerImpl manager) {
-			// Step 1: Find the set of shared ids
-			List<ModLoadOption> mandatoryMods = new ArrayList<>();
-			for (LoadOption option : duplicates) {
-				if (graph.isMandatory(option) || graph.isDepended(option)) {
-					mandatoryMods.add((ModLoadOption) option);
-				}
-			}
-
-			if (mandatoryMods.isEmpty()) {
-				// So this means there's an OptionalModIdDefinition with only
-				// DisabledModIdDefinitions as roots
-				// that means this isn't a duplicate mandatory mods error!
-				return;
-			}
-
-			// Try not to use a providing mod name
-			Iterator<ModLoadOption> iterator = mandatoryMods.iterator();
-			ModLoadOption firstMandatory = iterator.next();
-			while (graph.isProvided(firstMandatory) && iterator.hasNext()) {
-				firstMandatory = iterator.next();
-			}
-			String bestName = firstMandatory.metadata().name();
-
-			boolean sameVersion = duplicates.stream()
-					.filter(ModLoadOption.class::isInstance)
-					.map(ModLoadOption.class::cast)
-					.map(ModLoadOption::version)
-					.distinct()
-					.count() == 1;
-
-			// Title:
-			// Duplicate mod: "BuildCraft"
-
-			// Description:
-			// - "buildcraft-all-9.0.0.jar"
-			// - "buildcraft-all-9.0.1.jar"
-			// Remove all but one.
-
-			// With buttons to view each mod individually
-
-			QuiltLoaderText title = QuiltLoaderText.translate("error.duplicate_mandatory", bestName);
-			QuiltDisplayedError error = manager.theQuiltPluginContext.reportError(title);
-			error.appendReportText("'" + bestName + "'" + (!bestName.equals(id) ? (" ('" + id + "')") : "") + " is provided by multiple files:");
-			setIconFromMod(error, firstMandatory);
-
-			List<LoadOption> duplicates = new ArrayList<>(this.duplicates);
-			duplicates.sort((o1, o2) -> {
-				if (o1 instanceof ProvidedModOption) {
-					if (!(o2 instanceof ProvidedModOption)) {
-						return 1;
-					}
-				} else {
-					if (o2 instanceof ProvidedModOption) {
-						return -1;
-					}
-				}
-				if (o1 instanceof ModLoadOption && o2 instanceof ModLoadOption) {
-					int comp = ((ModLoadOption) o1).metadata().name().compareTo(((ModLoadOption) o2).metadata().name());
-					if (comp == 0) {
-						comp = ((ModLoadOption) o1).version().compareTo(((ModLoadOption) o2).version());
-					}
-
-					return comp;
-				}
-
-				return LoadOption.COMPARATOR.compare(o1, o2);
-			});
-
-			for (LoadOption loadOption : duplicates) {
-				if (loadOption instanceof ModLoadOption) {
-					ModLoadOption option = ((ModLoadOption) loadOption);
-					String path = manager.describePath(option.from());
-					Optional<Path> container = manager.getRealContainingFile(option.from());
-
-					// Just in case
-					if (option instanceof ProvidedModOption) {
-						error.appendDescription(QuiltLoaderText.translate("error.duplicate_mandatory.mod.provided", path));
-					} else {
-						error.appendDescription(QuiltLoaderText.translate("error.duplicate_mandatory.mod", path));
-					}
-
-					container.ifPresent(value -> error.addFileViewButton(QuiltLoaderText.translate("button.view_file", value.getFileName()), value)
-							.icon(option.modCompleteIcon()));
-
-					getModReportLine(option, manager, !option.metadata().name().equals(bestName), !sameVersion, true).forEach(error::appendReportText);
-					if (!graph.isMandatory(option)) {
-						if (graph.isDepended(option)) {
-							error.appendReportText("  - '" + option.metadata().name() + "' is needed by another mod, see Error <placeholder>");
-						} else {
-							// I think its only that this is an unless?
-							error.appendReportText("  - '" + option.metadata().name() + "' helps to resolve two or more mods that break each other, see Error <placeholder>");
-						}
-					}
-				} else {
-					error.appendDescription(QuiltLoaderText.translate("error.unhandled_mod_file.title", loadOption.describe()));
-					error.appendReportText("- Unknown load option: " + loadOption.describe());
-				}
-			}
-
-			error.appendDescription(QuiltLoaderText.translate("error.duplicate_mandatory.desc"));
-		}
 	}
 }
