+++ src/main/java/org/quiltmc/loader/impl/plugin/solvererror/graph/RootLink.java
@@ -1,0 +1,7 @@
+package org.quiltmc.loader.impl.plugin.solvererror.graph;
+
+/**
+ * Represents a link that does not represent a link between two options.
+ */
+public interface RootLink extends Link {
+}
